Show separate section for popular languages
This is helpful for quickly filtering by the most
common languages. I decided to include all
languages (including the most popular ones)
in the second `FilterCard`, because it was much
easier to use during testing if one didn't have to
switch between popular and "other" languages.
Instead, everything is in one list called
"all languages" below the top ones.